backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
openai>=1.40.0
pydantic>=2.7.0
python-dotenv>=1.0.0
beautifulsoup4
lxml
urllib3